sdk: actually turn it into a sdk
Move from `java_library` to `java_sdk_library` and expose
the packages we want to be usable.

Explicitly disable `unsafe_ignore_missing_latest_api` as this
would need some additional work.

For now this is enough to get started, catch changes to the
API and lint it as well.

To update the API:
- m update-api

or alternatively

- m org.lineageos.platform.stubs.source-update-current-api
- m org.lineageos.platform.stubs.source.system-update-current-api
- m org.lineageos.platform.stubs.source.test-update-current-api

For more information about java_sdk_library, see:
  - https://source.android.com/docs/setup/build/java-library
    - Be quick before they start moving the docs to
      "internal only" as well.

Change-Id: I8cbd5e8bf4c69fbe3e7685debd05686df96e8ed6
Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>